requirements.txt
sphinx-argparse
sphinx_rtd_theme
morfessor
pandas
numpy
GitPython
